+++ akshare/__init__.py
@@ -2516,9 +2516,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.10.99 fix: fix crypto_bitcoin_cme interface
 1.11.1 add: add index_us_stock_sina interface
 1.11.2 fix: fix stock_a_below_net_asset_statistics interface
+1.11.3 fix: fix stock_a_high_low_statistics interface
 """
 
-__version__ = "1.11.2"
+__version__ = "1.11.3"
 __author__ = "AKFamily"
 
 import sys