+++ docs/changelog.md
@@ -63,6 +63,10 @@
 
 ## 更新说明详情
 
+1.11.3 fix: fix stock_a_high_low_statistics interface
+
+    1. 修复 stock_a_high_low_statistics 接口
+
 1.11.2 fix: fix stock_a_below_net_asset_statistics interface
 
     1. 修复 stock_a_below_net_asset_statistics 接口
@@ -2836,6 +2840,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.11.3 fix: fix stock_a_high_low_statistics interface
 
 1.11.2 fix: fix stock_a_below_net_asset_statistics interface